requirements.txt
typing_extensions
langgraph
langchain-openai
langchain-core
langchain-community
langgraph-checkpoint-mongodb
python-dotenv
yfinance
pymongo
wikipedia
tmdbv3api
pandas
chromadb
openai
tmdbv3api==1.7.7